process/requirements.txt
Django
Django-ninja